Fixed issue wherein update processor would break on null index values

+++ src/main/java/com/pardot/rhombus/UpdateProcessor.java
@@ -122,14 +122,19 @@ public class UpdateProcessor {
 				listOfValuesToDelete.add(update);
 				listOfIndexesToDelete.addAll(
 					getListOfEffectedIndexes(objectMapper.getKeyspaceDefinition().getDefinitions().get(row.getObjectName()),
-					mostRecentUpdate,
-					update));
+						mostRecentUpdate,
+						update
+					)
+				);
 			}
 		}
 		//delete the list of indexes with a timestamp of the current update
 		for(CIndex index : listOfIndexesToDelete){
 			for(Map<String,Object> values: listOfValuesToDelete){
-				objectMapper.deleteObsoleteIndex(row,index, values);
+				//if any of the values are null we can skip it because its not actually a written index
+				if(!areAnyValuesNull(values)){
+					objectMapper.deleteObsoleteIndex(row,index, values);
+				}
 			}
 		}
 
@@ -138,6 +143,16 @@ public class UpdateProcessor {
 			objectMapper.deleteObsoleteUpdateIndexColumn(row.getRowKey(),todelete);
 		}
 
+	}
+
+	//todo add unit test
+	protected boolean areAnyValuesNull(Map<String,Object> indexValues){
+		for(String key : indexValues.keySet()){
+			if(indexValues.get(key) == null){
+				return true;
+			}
+		}
+		return false;
 	}
 
 	//todo add unit test
